# Add RootCache performance test