**Motivation**

RootCache may become unnecessary after merging SSZ v2.

- From https://github.com/ChainSafe/lodestar/issues/3998

**Description**

- Add performance test for RootCache
- Move RootCache out of registerAttestationInBlock

Locally I get

```
  RootCache.getBlockRootAtSlot
    ✔ RootCache.getBlockRootAtSlot - 250000 vs - 7PWei                 1.164144e+8 ops/s    8.590000 ns/op        -     573120 runs   7.15 s
    ✔ state getBlockRootAtSlot - 250000 vs - 7PWei                        974725.4 ops/s    1.025930 us/op        -       7795 runs   1.47 s
```

RootCache is faster (always will be since it's just a Map.get). But state is somewhat fast too. `1us` * 128 attestations per block is `0.13 ms` of processing time that would be added if we remove the RootCache. @tuyennhv what do you think?