@tuyennhv 2022-08-09
> RootCache is faster (always will be since it's just a Map.get). But state is somewhat fast too. 1us * 128 attestations per block is 0.13 ms of processing time that would be added if we remove the RootCache. @tuyennhv what do you think?

@dapplion I think ssz-v2 should have same mechanism to cache loaded nodes, do you know why it takes more time than RootCache? I'd drop it only when the performances between the two are comparable