@dapplion 2022-08-09
> I think ssz-v2 should have same mechanism to cache loaded nodes, do you know why it takes more time than RootCache?

To get a root each impl has to:
- ssz v1: traverse the state tree, traverse blockRoots tree, deserialize node to root
- ssz v2: use cache to jump to blockRoots, jump to root node, deserialize node to root
- RootCache: single Map.get()

RootCache will **always** be faster, since it doesn't have to deserialize anything. I don't think it's ever possible for the ssz impl to be as fast as the RootCache. The question is if the optimization is necessary having a RootCache there.