@tuyennhv 2022-08-10
> > I think ssz-v2 should have same mechanism to cache loaded nodes, do you know why it takes more time than RootCache?
> 
> To get a root each impl has to:
> 
> * ssz v1: traverse the state tree, traverse blockRoots tree, deserialize node to root
> * ssz v2: use cache to jump directly to root node, deserialize node to root
> * RootCache: single Map.get()
> 
> RootCache will **always** be faster, since it doesn't have to deserialize anything. I don't think it's ever possible for the ssz impl to be as fast as the RootCache. The question is if the optimization justifies having a RootCache there.

I see. Tbh, I still don't feel convenient for any changes that's slower 😃 
- RootCache is also used for validatorMonitor metric
- This is all about calling `hashObjectToByteArray()` 128 times (and with validatorMonitor, it's 256) for the same data in same flow vs calling it once and cache it, I'd go with the 2nd